refactor: move getExtensionFromAudioBlob to utils
