+++ apps/app/src/lib/services/DownloadServiceDesktopLive.ts
@@ -1,4 +1,5 @@
 import { DownloadService } from '$lib/services/DownloadService';
+import { getExtensionFromAudioBlob } from '$lib/utils';
 import { WhisperingError } from '@repo/shared';
 import { save } from '@tauri-apps/api/dialog';
 import { writeBinaryFile } from '@tauri-apps/api/fs';
@@ -27,15 +28,3 @@ export const DownloadServiceDesktopLive = Layer.succeed(
 			}),
 	}),
 );
-
-function getExtensionFromAudioBlob(blob: Blob) {
-	const mimeType = blob.type.toLowerCase();
-	const mimeIncludes = (...types: string[]) => types.some((type) => mimeType.includes(type));
-	if (mimeIncludes('webm')) return 'webm';
-	if (mimeIncludes('mp4', 'mpeg', 'mp4a')) return 'mp4';
-	if (mimeIncludes('ogg', 'opus')) return 'ogg';
-	if (mimeIncludes('wav', 'wave')) return 'wav';
-	if (mimeIncludes('aac')) return 'aac';
-	if (mimeIncludes('flac')) return 'flac';
-	return 'mp3';
-}
