+++ apps/app/src/lib/utils.ts
@@ -54,3 +54,15 @@ export const flyAndScale = (
 		easing: cubicOut,
 	};
 };
+
+export function getExtensionFromAudioBlob(blob: Blob) {
+	const mimeType = blob.type.toLowerCase();
+	const mimeIncludes = (...types: string[]) => types.some((type) => mimeType.includes(type));
+	if (mimeIncludes('webm')) return 'webm';
+	if (mimeIncludes('mp4', 'mpeg', 'mp4a')) return 'mp4';
+	if (mimeIncludes('ogg', 'opus')) return 'ogg';
+	if (mimeIncludes('wav', 'wave')) return 'wav';
+	if (mimeIncludes('aac')) return 'aac';
+	if (mimeIncludes('flac')) return 'flac';
+	return 'mp3';
+}
